· Fixed infinite recursion bug

+++ SceneR.Framework/Headers/System/Core.hpp
@@ -20,7 +20,7 @@ namespace System
     using UInt64  = uint64_t;
     using Single  = float;
     using Double  = double;
-    using Size    = uint32_t;
+    using Size    = size_t;
     using String  = std::u16string;
 }
